services/payment-service/requirements.txt
kafka-python
pydantic
sqlalchemy
databases
asyncpg
alembic